fix(throttle): Less finger-printable request delay
Following [discussion about the previous preset
delay](https://discord.com/channels/994270357957648404/994270357957648408/1405705039649046589),
we weren't able to find a reason for it and there was agreement that the example value
from yt-dlp is more sensible.

While we're at it, also set a default `min:` value that would probably be sensible if it
were to be used in the future.

Finally, capture the reasoning behind the new preset values in comments to make the
reference documentation more helpful to new users in the future.