build(docs): Running test_docgen reverts changes
Running `$ REGENERATE_DOCS=1 pytest tests/unit/docgen/test_docgen.py` reverts
uncommitted changes. Is this still necessary? If not, it would be best to remove it as
it increases the inner-loop iteration time when working on the docs. If it is still
necessary, why? Can it be changed to not revert uncommitted changes?